refactor(arika): split epoch.rs into submodules
epoch.rs had grown to 1448 lines mixing scale definitions, the leap-second
table, calendar conversion, scale conversions, and a 660-line test module.
Split it into a directory module (behavior-preserving; public API unchanged
via re-exports):

- epoch/mod.rs      — Epoch<S> struct, generic accessors, UTC constructors,
                      gmst/add/to_datetime, constants, re-exports
- epoch/scale.rs    — sealed TimeScale trait + Utc/Tai/Tt/Tdb/Ut1 markers
- epoch/leap.rs     — IERS LEAP_SECONDS table + tai_minus_utc_at_mjd
- epoch/datetime.rs — DateTime + Display + Julian-Date -> calendar conversion
- epoch/duration.rs — Duration
- epoch/convert.rs  — scale conversions, scale-specific JD constructors,
                      era_formula, tdb_minus_tt
- epoch/tests.rs    — the unit test module, verbatim

The split follows the 'coarse + convert.rs' shape: the most tightly coupled
and bug-prone part (scale conversions, routed through TAI) is isolated, while
the UTC constructors stay with Epoch for discoverability. Leap-second
convergence loops, the UTC->TAI->TT->TDB path, and the ERA constant are moved
verbatim — no numeric changes.

from_jd_raw stays pub(crate); helpers shared across submodules are pub(super).
The null_eop_in_to_ut1 trybuild fixture is updated: to_ut1 moved files, so the
"required by a bound" note now points at epoch/convert.rs (the compile error
itself — NullEop rejected — is unchanged).

Verified: arika tests (54 epoch + trybuild), clippy (std/no_std/alloc),
fmt, rustdoc (epoch links clean), wasm32 (arika no_std+alloc + arika-wasm),
and downstream orts builds against the re-exported API.

Refs #104
